GyanGuru/requirements.txt
Flask
python-dotenv
google-generativeai
gTTS
Pillow